+++ src/main/java/com/avp/common/block/entity/DeskTerminalBlockEntity.java
@@ -4,9 +4,9 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 
-public class DeskTerminalBE extends BlockEntity {
+public class DeskTerminalBlockEntity extends BlockEntity {
 
-    public DeskTerminalBE(BlockPos pos, BlockState blockState) {
-        super(BlockEntityTypes.DESK_TERMINAL_BE, pos, blockState);
+    public DeskTerminalBlockEntity(BlockPos pos, BlockState blockState) {
+        super(AVPBlockEntityTypes.DESK_TERMINAL, pos, blockState);
     }
 }