+++ src/main/java/com/avp/common/block/entity/LeadChestBlockEntity.java
@@ -9,22 +9,22 @@ import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.Container;
 import net.minecraft.world.ContainerHelper;
-import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.inventory.ChestMenu;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.entity.*;
+import net.minecraft.world.level.block.entity.BaseContainerBlockEntity;
+import net.minecraft.world.level.block.entity.ChestLidController;
+import net.minecraft.world.level.block.entity.ContainerOpenersCounter;
+import net.minecraft.world.level.block.entity.LidBlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import org.jetbrains.annotations.NotNull;
 
 import com.avp.common.block.AVPBlocks;
-import com.avp.common.item.AVPItemTags;
-import com.avp.common.item.AVPItems;
 
-public class AmmoChestBE extends RandomizableContainerBlockEntity implements LidBlockEntity {
+public class LeadChestBlockEntity extends BaseContainerBlockEntity implements LidBlockEntity {
 
     private final ChestLidController chestLidController = new ChestLidController();
 
@@ -32,8 +32,8 @@ public class AmmoChestBE extends RandomizableContainerBlockEntity implements Lid
 
     private NonNullList<ItemStack> itemStacks = NonNullList.withSize(27, ItemStack.EMPTY);
 
-    public AmmoChestBE(BlockPos pos, BlockState blockState) {
-        super(BlockEntityTypes.AMMO_CHEST_BE, pos, blockState);
+    public LeadChestBlockEntity(BlockPos pos, BlockState blockState) {
+        super(AVPBlockEntityTypes.LEAD_CHEST, pos, blockState);
 
         this.openersCounter = new ContainerOpenersCounter() {
 
@@ -67,17 +67,15 @@ public class AmmoChestBE extends RandomizableContainerBlockEntity implements Lid
 
             @Override
             protected void openerCountChanged(Level level, BlockPos pos, BlockState state, int count, int openCount) {
-                level.blockEvent(AmmoChestBE.this.worldPosition, AVPBlocks.AMMO_CHEST, 1, openCount);
+                level.blockEvent(LeadChestBlockEntity.this.worldPosition, AVPBlocks.LEAD_CHEST, 1, openCount);
             }
 
             @Override
             protected boolean isOwnContainer(Player player) {
-                if (!(player.containerMenu instanceof ChestMenu)) {
+                if (!(player.containerMenu instanceof ChestMenu))
                     return false;
-                }
-
                 Container container = ((ChestMenu) player.containerMenu).getContainer();
-                return container == AmmoChestBE.this;
+                return container == LeadChestBlockEntity.this;
             }
         };
     }
@@ -85,24 +83,18 @@ public class AmmoChestBE extends RandomizableContainerBlockEntity implements Lid
     @Override
     protected void saveAdditional(CompoundTag tag, HolderLookup.Provider registries) {
         super.saveAdditional(tag, registries);
-
-        if (!this.trySaveLootTable(tag)) {
-            ContainerHelper.saveAllItems(tag, this.itemStacks, registries);
-        }
+        ContainerHelper.saveAllItems(tag, this.itemStacks, false, registries);
     }
 
     @Override
     protected void loadAdditional(CompoundTag tag, HolderLookup.Provider registries) {
         super.loadAdditional(tag, registries);
         this.itemStacks = NonNullList.withSize(this.getContainerSize(), ItemStack.EMPTY);
-
-        if (!this.tryLoadLootTable(tag)) {
+        if (tag.contains("Items", 9))
             ContainerHelper.loadAllItems(tag, this.itemStacks, registries);
-        }
     }
 
-    @SuppressWarnings("unused")
-    public static void lidAnimateTick(Level level, BlockPos pos, BlockState state, AmmoChestBE blockEntity) {
+    public static void lidAnimateTick(Level level, BlockPos pos, BlockState state, LeadChestBlockEntity blockEntity) {
         blockEntity.chestLidController.tickLid();
     }
 
@@ -116,9 +108,9 @@ public class AmmoChestBE extends RandomizableContainerBlockEntity implements Lid
         if (id == 1) {
             this.chestLidController.shouldBeOpen(type > 0);
             return true;
+        } else {
+            return super.triggerEvent(id, type);
         }
-
-        return super.triggerEvent(id, type);
     }
 
     @Override
@@ -143,7 +135,7 @@ public class AmmoChestBE extends RandomizableContainerBlockEntity implements Lid
 
     @Override
     protected @NotNull Component getDefaultName() {
-        return Component.translatable("container.ammo_chest");
+        return Component.translatable("container.lead_chest");
     }
 
     @Override
@@ -157,36 +149,6 @@ public class AmmoChestBE extends RandomizableContainerBlockEntity implements Lid
     }
 
     @Override
-    public void setItem(int slot, ItemStack stack) {
-        if (isValidForChest(stack)) {
-            super.setItem(slot, stack);
-            this.setChanged();
-        } else {
-            if (!stack.isEmpty() && this.level != null) {
-                this.dropItem(stack);
-            }
-        }
-    }
-
-    private void dropItem(ItemStack stack) {
-        if (this.level != null && !this.level.isClientSide) {
-            this.level.addFreshEntity(
-                new ItemEntity(
-                    this.level,
-                    this.worldPosition.getX() + 0.5,
-                    this.worldPosition.getY() + 0.5,
-                    this.worldPosition.getZ() + 0.5,
-                    stack
-                )
-            );
-        }
-    }
-
-    private boolean isValidForChest(ItemStack stack) {
-        return stack.isEmpty() || stack.is(AVPItemTags.AMMO_ITEMS);
-    }
-
-    @Override
     protected @NotNull AbstractContainerMenu createMenu(int containerId, Inventory inventory) {
         return ChestMenu.threeRows(containerId, inventory, this);
     }
@@ -194,23 +156,5 @@ public class AmmoChestBE extends RandomizableContainerBlockEntity implements Lid
     @Override
     public int getContainerSize() {
         return this.itemStacks.size();
-    }
-
-    public boolean hasAmmo() {
-        return !this.itemStacks.isEmpty() && this.itemStacks.stream().anyMatch(item -> item.is(AVPItems.MEDIUM_BULLET));
-    }
-
-    public boolean consumeAmmo(int count) {
-        for (ItemStack itemStack : this.itemStacks) {
-            if (itemStack.is(AVPItems.MEDIUM_BULLET)) {
-                var available = itemStack.getCount();
-                if (available >= count) {
-                    itemStack.shrink(count);
-                    this.setChanged();
-                    return true;
-                }
-            }
-        }
-        return false;
     }
 }